Moved discordgo wrapper related functions to their own file
