+++ sds.go
@@ -13,7 +13,6 @@ import (
 	"io/ioutil"
 	"math/rand"
 	"os"
-	"os/exec"
 	"os/signal"
 	"strings"
 	"syscall"
@@ -79,6 +78,12 @@ func main() {
 		sendSDSMsg(&isFirstSDSTime, guildMap, reverseGuildMap, totalGuilds, dg)
 	})
 
+	/* Schedule a job to update the listening rich presense every 3 mins, kinf
+	of acts as a heartbeat */
+	scheduler.Every(3).Minutes().Run(func() {
+		updateListening(dg)
+	})
+
 	// Wait here until CTRL-C is recieved
 	fmt.Println("[INFO] SDS is now running. Press CTRL-C to exit.")
 	sc := make(chan os.Signal, 1)
@@ -89,13 +94,6 @@ func main() {
 	dg.Close()
 	writeMsgsToFile(guildMap, reverseGuildMap, &totalGuilds)
 	fmt.Println("\n[INFO] Bot has successfully closed. Goodnight sweet prince")
-}
-
-// Ran when recieves the "ready" status from Discord
-func ready(s *discordgo.Session, event *discordgo.Ready) {
-
-	// Set the playing status.
-	updateListening(s)
 }
 
 func updateListening(s *discordgo.Session) {
@@ -190,47 +188,6 @@ func sendSDSMsg(isFirstTime *bool, guildMap map[string]guildData, reverseGuildMa
 			if c.Name == "general" {
 				s.ChannelMessageSend(c.ID, string(msg)[1:])
 			}
-		}
-	}
-}
-
-// Function called every time a new message is created in a bot-authorized chan
-func messageCreate(s *discordgo.Session, m *discordgo.MessageCreate) {
-	// Ignore all messages created by the bot itself
-	if m.Author.ID == s.State.User.ID {
-		return
-	}
-
-	// If user requests for information about the server
-	if strings.HasPrefix(m.Content, "./machine") {
-		// Run neofetch command and get output
-		out, err := exec.Command("neofetch", "--disable", "underline",
-			"--stdout").Output()
-
-		// Error check and make sure everything happened correctly
-		if err != nil {
-			fmt.Println("[ERR!] Could not run ./machine command")
-			return
-		}
-
-		// Convert neofetch output to string
-		output := string(out)
-
-		// Send the message
-		s.ChannelMessageSend(m.ChannelID, "```"+output+"```")
-	} else {
-		// Check if the message is blank, if it is then do not save it
-		if m.Content == "" {
-			fmt.Println("[INFO] Blank message sent, nothing will be logged")
-			return
-		}
-
-		// Otherwise log the message
-		newMsg := discordMessage{m.Content, m.GuildID}
-		msgQueue = append(msgQueue, newMsg)
-		fmt.Println("Current queue: ")
-		for _, msgs := range msgQueue {
-			fmt.Println("[" + msgs.guild + " : " + msgs.msg + "]")
 		}
 	}
 }
